reorganize Tabs
