+++ src/components/Tabs/Tabs.tsx
@@ -1,11 +1,6 @@
 import { FC, useMemo, useState } from 'react';
-import cx from 'classnames';
 
-interface SharedTabProps {
-  borderClassName?: string;
-  Icon: FC;
-  rounded?: 'full' | 't-md';
-}
+import { SharedTabProps, Tab } from './Tab';
 
 export type TabProps = SharedTabProps & { content: FC };
 
@@ -42,29 +37,3 @@ export const Tabs: FC<TabsProps> = ({ tabs, children, tabIndex = 0 }) => {
     </>
   );
 };
-
-type IProps = SharedTabProps & {
-  active?: boolean;
-  onClick: () => void;
-}
-
-export const Tab: FC<IProps> = ({
-  active = false,
-  borderClassName: color,
-  Icon,
-  onClick,
-  rounded = 't-md'
-}) => (
-  <div
-    className={cx(
-      'p-3 hover:bg-gray-200 cursor-pointer',
-      color,
-      rounded === 't-md' && 'rounded-t-md',
-      rounded === 'full' && 'rounded-full',
-      active && 'border-b-2'
-    )}
-    onClick={onClick}
-  >
-    <Icon />
-  </div>
-);
